fix(linux): use rigaya images and package nvJPEG2000

+++ build_rpm.sh
@@ -1,5 +1,7 @@
 #!/bin/sh
 
+set -e
+
 PACKAGE_NAME=nvencc
 PACKAGE_BIN=nvencc
 PACKAGE_OS=
@@ -9,7 +11,7 @@ PACKAGE_DESCRIPTION=
 PACKAGE_ROOT=.rpmpkg
 PACKAGE_VERSION=`git describe --tags | cut -f 1 --delim="-" | sed -e 's/^v//'`
 PACKAGE_ARCH=`uname -m`
-PACKAGE_LICENSE=MIT
+PACKAGE_LICENSE="MIT AND LicenseRef-NVIDIA-nvJPEG2000"
 
 if [ -e /etc/os-release ]; then
     PACKAGE_OS_NAME=`cat /etc/os-release | grep ID | cut -f 2 --delim="=" | head -n 1 `
@@ -17,9 +19,22 @@ if [ -e /etc/os-release ]; then
     PACKAGE_OS="${PACKAGE_OS_NAME}${PACKAGE_OS_VER}"
 fi
 
-mkdir -p ${PACKAGE_ROOT}
-cp ${PACKAGE_BIN} ${PACKAGE_ROOT}
-chmod +x ${PACKAGE_ROOT}/${PACKAGE_BIN}
+NVJPEG2K_LIB_DIR=${NVJPEG2K_ROOT:?NVJPEG2K_ROOT is not set}/lib/${NVJPEG2K_CUDA_MAJOR:?NVJPEG2K_CUDA_MAJOR is not set}
+if [ ! -f "${NVJPEG2K_LIB_DIR}/libnvjpeg2k.so.0.9.0.43" ] || [ ! -L "${NVJPEG2K_LIB_DIR}/libnvjpeg2k.so" ] || [ ! -f "${NVJPEG2K_ROOT}/LICENSE" ]; then
+    echo "nvJPEG2000 runtime or license is missing under ${NVJPEG2K_ROOT}!"
+    exit 1
+fi
+
+rm -rf ${PACKAGE_ROOT}
+mkdir -p ${PACKAGE_ROOT}/usr/bin
+cp ${PACKAGE_BIN} ${PACKAGE_ROOT}/usr/bin/
+chmod +x ${PACKAGE_ROOT}/usr/bin/${PACKAGE_BIN}
+
+mkdir -p ${PACKAGE_ROOT}/usr/lib64
+cp -a ${NVJPEG2K_LIB_DIR}/libnvjpeg2k.so* ${PACKAGE_ROOT}/usr/lib64/
+
+mkdir -p ${PACKAGE_ROOT}/usr/share/licenses/${PACKAGE_NAME}
+cp ${NVJPEG2K_ROOT}/LICENSE ${PACKAGE_ROOT}/usr/share/licenses/${PACKAGE_NAME}/nvjpeg2000-LICENSE
 
 rm -rf rpmbuild
 RPMBUILD_DIR=${PWD}/rpmbuild
@@ -35,7 +50,7 @@ mkdir ${WORK_DIR}
 
 build_pkg/replace.py \
     --rpm \
-    -i build_pkg/template.spec \
+    -i packaging/nvencc.spec \
     -o ${WORK_DIR}/${PACKAGE_NAME}.spec \
     --pkg-name ${PACKAGE_NAME} \
     --pkg-bin ${PACKAGE_BIN} \
@@ -44,7 +59,7 @@ build_pkg/replace.py \
     --pkg-maintainer ${PACKAGE_MAINTAINER} \
     --pkg-depends ${PACKAGE_DEPENDS} \
     --pkg-desc ${PACKAGE_DESCRIPTION} \
-    --pkg-license ${PACKAGE_LICENSE}
+    --pkg-license "${PACKAGE_LICENSE}"
 
 cp -rp "${PACKAGE_ROOT}" "${WORK_DIR}/${PACKAGE_NAME}"
 cd ${WORK_DIR}
